Multiple changes

+++ pdfcmd.py
@@ -5,25 +5,6 @@ import os
 import logging
 import sys
 
-logging.basicConfig(
-    level=logging.INFO,
-    format="%(levelname)-10s: %(message)s",
-    handlers=[logging.StreamHandler()]
-)
-
-parser = OptionParser("Script to modify pdf-files (rotate, extract, delete pages, etc.).")
-parser.add_option("-f", "--files",      action="append",    type="string",  dest="filenames",   help="Filepaths of files to be modified.")
-parser.add_option("-o", "--open",       action="store_true",                dest="open",        help="#TODO")
-
-# TODO Group
-#TODO: argparse statt optionparser
-parser.add_option("-r", "--rotate",     action="store",     type="string",  dest="rotate",      help="Direction and Page-numbers to be rotated (e.g. for 90°-clockwise 'R2-4,5').")
-parser.add_option("-d", "--delete",     action="store",     type="string",  dest="delete",      help="Page-numbers to be deleted (e.g. '2-4,5').")
-parser.add_option("-m", "--merge",      action="store",     type="string",  dest="merge",       help="Output-Path of merged-pdf. All 'files' will be merged into one.")
-parser.add_option("-e", "--extract",    action="store",     type="string",  dest="extract",     help="Page-numbers to be extracted (e.g. '2-4,5').")
-parser.add_option("-s", "--split",      action="store",     type="string",  dest="split",       help="Page-numbers after initial pdf will be splitted (e.g. for '2', pdf will be split between 2nd and 3rd page).")
-options, args = parser.parse_args()
-
 
 def main():
     # READ -----------------------------------------------------------------------------------------------------------
@@ -35,7 +16,7 @@ def main():
             os.system(f'start chrome "{os.path.abspath(filename)}"')
 
     del_src = False
-    save_src = False
+    save_src = False if not options.remove_password else True
 
     # 1 ROTATE ---------------------------------------------------------------------------------------------------------
     if options.rotate is not None:
@@ -82,14 +63,19 @@ def main():
     if save_src:
         logging.info("Saving...")
         for filename, pdf in zip(filenames, pdfs):
+            if options.interactive and input(f"Save/Overwrite '{filename}'? [y/n] ") != "y":
+                continue
             logging.info(f"\t{filename}")
             save((pdf, filename))
+            if options.open:
+                os.system(f'start chrome "{os.path.abspath(filename)}"')
     if del_src:
         logging.info("Deleting initial file(s)...")
         for filename in filenames:
+            if options.interactive and input(f"Delete '{filename}'? [y/n] ") != "y":
+                continue
             logging.info(f"\t{filename}")
             os.remove(filename)
-    # TODO: kombinationen ausschließen -> nicht seiten deleten und dann extracten
 
 
 def read() -> tuple:
@@ -103,7 +89,7 @@ def read() -> tuple:
         if os.path.exists(filename):
             filenames.append(filename)
         else:
-            tmp_filenames = glob(filename)
+            tmp_filenames = sorted(glob(filename))
             filenames += tmp_filenames
             if len(tmp_filenames) == 0:
                 logging.warning(f"\tNo files found for '{filename}'")
@@ -111,7 +97,13 @@ def read() -> tuple:
     # Read pdfs
     pdfs = []
     for filename in filenames:
-        pdfs.append(pikepdf.Pdf.open(filename, allow_overwriting_input=True))
+        if options.interactive and input(f"Read '{filename}'? [y/n] ") != "y":
+            continue
+
+        if options.password is None:
+            pdfs.append(pikepdf.Pdf.open(filename, allow_overwriting_input=True))
+        else:
+            pdfs.append(pikepdf.Pdf.open(filename, allow_overwriting_input=True, password=options.password))
     return pdfs, filenames
 
 
@@ -129,7 +121,6 @@ def rotate(pdf):
     """
     Syntax:
     Rotate left = Rotate counterclockwiese = L
-    #TODO R2n und V2n+1
     :param pdf:
     :return:
     """
@@ -152,7 +143,7 @@ def rotate(pdf):
         if seq.strip().isdigit() and 0 <= int(seq.strip())-1 < n:
             all_idx[int(seq)-1] = True
         elif "n" in seq:
-            pass  # TODO
+            raise NotImplementedError  # TODO
         elif "-" in seq:
             lower, upper = [int(digit.strip()) for digit in seq.split("-")]
             all_idx[lower-1:upper] = [True for _ in range(lower-1, upper)]
@@ -180,7 +171,6 @@ def rotate(pdf):
 
 def delete(pdf):
     """
-    #TODO
     :param pdf:
     :return:
     """
@@ -193,7 +183,7 @@ def delete(pdf):
         if seq.isdigit() and 0 <= int(seq) - 1 < n:
             all_idx[int(seq) - 1] = True
         elif "n" in seq:
-            pass  # TODO
+            raise NotImplementedError  # TODO
         elif len(seq.split("-")) == 2:
             lower, upper = [i for i in seq.split("-")]
             if upper == "" and lower != "":
@@ -215,7 +205,6 @@ def delete(pdf):
 
 def extract(pdf):
     """
-    # TODO
     :param pdf:
     :return:
     """
@@ -228,7 +217,7 @@ def extract(pdf):
         if seq.isdigit() and 0 <= int(seq) - 1 < n:
             all_idx[int(seq) - 1] = True
         elif "n" in seq:
-            pass  # TODO
+            raise NotImplementedError  # TODO
         elif len(seq.split("-")) == 2:
             lower, upper = [i for i in seq.split("-")]
             if upper == "" and lower != "":
@@ -305,5 +294,24 @@ def split(pdf):
 
 
 if __name__ == "__main__":
-    main()
+    logging.basicConfig(
+        level=logging.INFO,
+        format="%(levelname)-10s: %(message)s",
+        handlers=[logging.StreamHandler()]
+    )
 
+    parser = OptionParser("Script to modify pdf-files (rotate, extract, delete pages, etc.).")
+    parser.add_option("-f", "--files",              action="append",    type="string",  dest="filenames",       help="Filepaths of files to be modified.")
+    parser.add_option("-o", "--open",               action="store_true",                dest="open",            help="Opens files after process is done.")
+    parser.add_option("-i", "--interactive",        action="store_true",                dest="interactive",     help="Ask before each operation")
+
+    parser.add_option("-r", "--rotate",             action="store",     type="string",  dest="rotate",          help="Direction and Page-numbers to be rotated (e.g. for 90°-clockwise 'R2-4,5').")
+    parser.add_option("-d", "--delete",             action="store",     type="string",  dest="delete",          help="Page-numbers to be deleted (e.g. '2-4,5').")
+    parser.add_option("-m", "--merge",              action="store",     type="string",  dest="merge",           help="Output-Path of merged-pdf. All 'files' will be merged into one.")
+    parser.add_option("-e", "--extract",            action="store",     type="string",  dest="extract",         help="Page-numbers to be extracted (e.g. '2-4,5').")
+    parser.add_option("-s", "--split",              action="store",     type="string",  dest="split",           help="Page-numbers after initial pdf will be splitted (e.g. for '2', pdf will be split between 2nd and 3rd page).")
+    parser.add_option("-p", "--password",           action="store",     type="string",  dest="password",        help="Supply Password for protected files.")
+    parser.add_option(      "--remove-password",    action="store_true",                dest="remove_password", help="Remove password.")
+    options, args = parser.parse_args()
+
+    main()
